fix: split dogfood app theme dojo debt
