+++ scripts/code-dojo/baselines/eslint.json
@@ -1,12 +1,12 @@
 {
   "schema": "code-dojo.eslint-baseline.v1",
-  "total": 317,
-  "errors": 317,
+  "total": 266,
+  "errors": 266,
   "warnings": 0,
   "rules": [
     {
       "ruleId": "@typescript-eslint/consistent-type-assertions",
-      "count": 9
+      "count": 8
     },
     {
       "ruleId": "@typescript-eslint/no-base-to-string",
@@ -25,10 +25,6 @@
       "count": 2
     },
     {
-      "ruleId": "@typescript-eslint/no-empty-object-type",
-      "count": 1
-    },
-    {
       "ruleId": "@typescript-eslint/no-explicit-any",
       "count": 2
     },
@@ -38,7 +34,7 @@
     },
     {
       "ruleId": "@typescript-eslint/no-non-null-assertion",
-      "count": 66
+      "count": 46
     },
     {
       "ruleId": "@typescript-eslint/no-redundant-type-constituents",
@@ -46,7 +42,7 @@
     },
     {
       "ruleId": "@typescript-eslint/no-unnecessary-condition",
-      "count": 30
+      "count": 28
     },
     {
       "ruleId": "@typescript-eslint/no-unnecessary-type-arguments",
@@ -54,7 +50,7 @@
     },
     {
       "ruleId": "@typescript-eslint/no-unnecessary-type-assertion",
-      "count": 3
+      "count": 2
     },
     {
       "ruleId": "@typescript-eslint/no-unnecessary-type-conversion",
@@ -70,7 +66,7 @@
     },
     {
       "ruleId": "@typescript-eslint/no-unsafe-assignment",
-      "count": 5
+      "count": 4
     },
     {
       "ruleId": "@typescript-eslint/no-unsafe-member-access",
@@ -82,11 +78,11 @@
     },
     {
       "ruleId": "@typescript-eslint/no-unsafe-type-assertion",
-      "count": 67
+      "count": 60
     },
     {
       "ruleId": "@typescript-eslint/no-unused-vars",
-      "count": 6
+      "count": 5
     },
     {
       "ruleId": "@typescript-eslint/non-nullable-type-assertion-style",
@@ -101,10 +97,6 @@
       "count": 13
     },
     {
-      "ruleId": "@typescript-eslint/prefer-regexp-exec",
-      "count": 1
-    },
-    {
       "ruleId": "@typescript-eslint/require-await",
       "count": 2
     },
@@ -114,26 +106,18 @@
     },
     {
       "ruleId": "@typescript-eslint/restrict-template-expressions",
-      "count": 56
+      "count": 44
     },
     {
       "ruleId": "@typescript-eslint/unified-signatures",
-      "count": 3
+      "count": 2
     },
     {
       "ruleId": "no-control-regex",
       "count": 4
     },
     {
-      "ruleId": "no-restricted-imports",
-      "count": 2
-    },
-    {
       "ruleId": "no-useless-assignment",
-      "count": 1
-    },
-    {
-      "ruleId": "no-useless-escape",
       "count": 1
     }
   ]
